python-services/requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
spacy==3.7.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
requests==2.32.3
pytesseract==0.3.10
pillow==10.4.0
sentence-transformers==3.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
opencv-python==4.10.0.84
pdf2image==1.17.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
transformers==4.45.2
torch==2.4.1
accelerate==0.34.2
datasets==3.0.1
evaluate==0.4.3
nltk==3.9.1
textblob==0.18.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2